Add filename option to log formatting
The file option we use by default right now is very long and usually
includes local paths on the host the software was built on. This
modification provides a filename option to shorten the length of the
prefix for the output while still retaining the ability to find where
the log was coming from.

The default is updated to use the filename option as well.