Fix Python CI breakages (#262)
Mitigate drift in supported GitHub runner versions:
* ubuntu-latest now uses ubuntu-24.04, which broke our Python CI due to newer and stricter GCC.
* macos-12 is EOL.

Fix (1) by pinning Python builds to ubuntu-22.04. Note that ubuntu-24.04 also drops out-of-the-box support for Python 3.8, which we should continue testing to match the Brain.

Fix (2) by upgrading to macos-13, likewise the final runner with Python 3.8 baked in. This required a fix to an undefined behavior (compiler-dependent) bug in the pybindings for Lie group types.

Completes SWE-531 and SWE-538.